Moved the Standard NameConverter into its own class
